requirements.txt
blinker==1.6.3
click==8.1.7
colorama==0.4.6
Flask==3.0.0
imageio==2.31.5
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
lazy_loader==0.3
MarkupSafe==2.1.3
networkx==3.1
numpy==1.26.0
opencv-python-headless==4.8.1.78
packaging==23.2
Pillow==10.0.1
scikit-image==0.22.0
scikit-learn==1.3.1
scipy==1.11.3
threadpoolctl==3.2.0
tifffile==2023.9.26
Werkzeug==3.0.0
